EDU-363 Remove service to service dependencies involving storage

+++ src/server/storage-controller.js
@@ -37,7 +37,7 @@ class StorageController {
     const { prefix } = req.query;
     const { objectName } = req.params;
 
-    await this.storageService.deleteObject({ prefix, objectName, user });
+    await this.storageService.deleteObject({ prefix, objectName, userId: user._id });
 
     return res.send({});
   }
@@ -68,7 +68,7 @@ class StorageController {
       }
     }
 
-    await this.storageService.uploadFiles({ prefix, files, user });
+    await this.storageService.uploadFiles({ prefix, files, userId: user._id });
     return res.status(201).send({});
   }
